backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
pymongo==4.0.1
python-dotenv==0.19.2
pyjwt==2.3.0
werkzeug==2.0.2
redis==4.0.2
requests==2.28.1
waitress>=2.0